pyproject.toml
[project]
name = "full-time-equivalent-project"
version = "0.1.0"
description = "Personal AI Employee — Bronze Tier Foundation"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "watchdog>=4.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
